moving path_to_string to specs
